refactor project architecture
