--- pcl_apps/src/clustering/CMakeLists.txt
@@ -1,12 +1,0 @@
-# Euclidean Clustering Component/Node
-ament_auto_add_library(euclidean_clustering_component SHARED
-  euclidean_clustering/euclidean_clustering_component.cpp
-)
-target_compile_definitions(euclidean_clustering_component PRIVATE "PCL_APPS_EUCLIDEAN_CLUSTERING_BUILDING_DLL")
-target_link_libraries(euclidean_clustering_component
-  boost_system ${PCL_LIBRARIES})
-
-ament_auto_add_executable(euclidean_clustering_node
-  euclidean_clustering/euclidean_clustering_node.cpp)
-target_link_libraries(euclidean_clustering_node
-  euclidean_clustering_component ${PCL_LIBRARIES})
