--- pcl_apps/src/matching/CMakeLists.txt
@@ -1,27 +1,0 @@
-# Ndt Matching Component/Node
-ament_auto_add_library(ndt_matching_component SHARED
-  ndt_matching/ndt_matching_component.cpp
-)
-target_compile_definitions(ndt_matching_component PRIVATE "PCL_APPS_NDT_MATCHING_BUILDING_DLL")
-target_link_libraries(ndt_matching_component
-  ${PCL_LIBRARIES})
-
-ament_auto_add_executable(ndt_matching_node
-  ndt_matching/ndt_matching_node.cpp
-)
-target_link_libraries(ndt_matching_node
-  ndt_matching_component ${PCL_LIBRARIES})
-
-# Ndt Matching Twist Estimator Component/Node
-ament_auto_add_library(ndt_matching_twist_estimator_component SHARED
-  ndt_matching/ndt_matching_twist_estimator_component.cpp
-)
-target_compile_definitions(ndt_matching_twist_estimator_component PRIVATE "PCL_APPS_NDT_MATCHING_TWIST_ESTIMATOR_BUILDING_DLL")
-target_link_libraries(ndt_matching_twist_estimator_component
-  ${PCL_LIBRARIES})
-
-ament_auto_add_executable(ndt_matching_twist_estimator_node
-  ndt_matching/ndt_matching_twist_estimator_node.cpp
-)
-target_link_libraries(ndt_matching_twist_estimator_node
-ndt_matching_twist_estimator_component ${PCL_LIBRARIES})
